protect api routes with jwt verification from auth0 token

+++ src/app/api/collections/[id]/route.ts
@@ -3,6 +3,7 @@ import { getCollection } from "@/src/dal/collection/queries";
 import { GetCollectionResponse } from "@/src/lib/types";
 import { NextRequest, NextResponse } from "next/server";
 import { Prisma } from "@/src/generated/prisma/client";
+import { validateToken, hasPermissions } from "@/src/dal/utils";
 
 export async function GET(
   request: NextRequest,
@@ -40,6 +41,52 @@ export async function DELETE(
   { params }: { params: Promise<{ id: string }> },
 ) {
   try {
+    const authHeader = request.headers.get("Authorization");
+
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "No token provided",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const token = authHeader.split(" ")[1];
+    const authPayload = await validateToken(token);
+
+    if (!authPayload) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Invalid token",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const isAuthorized = hasPermissions(
+      authPayload.permissions as Array<string>,
+      "delete:collections",
+    );
+
+    if (!isAuthorized) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Unauthorized",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
     const deletedCollectionId = await deleteCollection((await params).id);
 
     const response = {
